Split edit profile screen into smaller components
